refactor: move `D` operator into `DynamicAutodiff` submodule
