+++ src/TemplateExpression.jl
@@ -28,6 +28,7 @@ using ..CoreModule:
     AbstractOptions, Options, Dataset, CoreModule as CM, AbstractMutationWeights, has_units
 using ..ConstantOptimizationModule: ConstantOptimizationModule as CO
 using ..InterfaceDynamicExpressionsModule: InterfaceDynamicExpressionsModule as IDE
+using ..DynamicAutodiffModule: DynamicAutodiffModule as DA
 using ..MutationFunctionsModule: MutationFunctionsModule as MF
 using ..ExpressionBuilderModule: ExpressionBuilderModule as EB
 using ..HallOfFameModule: HallOfFameModule as HOF
@@ -38,8 +39,6 @@ using ..LossFunctionsModule: LossFunctionsModule as LF
 using ..MutateModule: MutateModule as MM
 using ..PopMemberModule: PopMember
 using ..ComposableExpressionModule: ComposableExpression, ValidVector
-
-import ..ComposableExpressionModule: D
 
 """
     TemplateStructure{K,E,NF} <: Function
@@ -95,7 +94,7 @@ end
 
 # We pass through the derivative operators, since
 # we just want to record the number of arguments.
-D(f::ArgumentRecorder, _::Integer) = f
+DA.D(f::ArgumentRecorder, _::Integer) = f
 
 """Infers number of features used by each subexpression, by passing in test data."""
 function infer_variable_constraints(::Val{K}, combiner::F) where {K,F}
